Reference

SimplexPlots.simplex_plot — Function

simplex_plot(x::AbstractMatrix{Float64};
             vertices = ["x₁", "x₂", "x₃"],
             axes = [1, 2, 3],
             args...)

Plot vectors in x. Vertex labels can be set with the vertices keyword argument (set this to nothing to suppress vertex labels). Further keyword arguments are passed to RecipesBase.plot.

simplex_plot(x::Vector{Vector{Float64}};
             vertices = ["x₁", "x₂", "x₃"],
             axes = [1, 2, 3],
             args...)

simplex_plot(x::Vector{Float64};
             vertices = ["x₁", "x₂", "x₃"],
             axes = [1, 2, 3]
             args...)

SimplexPlots.simplex_plot! — Function

simplex_plot!(x::AbstractMatrix{Float64};
              axes = [1, 2, 3],
              args...)

Like simplex_plot, but adds to an existing plot.

simplex_plot!(x::Vector{Vector{Float64}};
              axes = [1, 2, 3],
              args...)

simplex_plot!(x::Vector{Float64};
              axes = [1, 2, 3],
              args...)

SimplexPlots.simplex_scatter — Function

simplex_scatter(x::AbstractMatrix{Float64};
                vertices = ["x₁", "x₂", "x₃"],
                axes = [1, 2, 3],
                args...)

Scatterplot of the vectors in x. Vertex labels can be set using the vertices keyword argument (set this to nothing to suppress vertex labels). Further keyword arguments are passed to Plots.scatter.

simplex_scatter(x::Vector{Vector{Float64}};
                vertices = ["x₁", "x₂", "x₃"],
                axes = [1, 2, 3],
                args...)

simplex_scatter(x::Vector{Float64};
                vertices = ["x₁", "x₂", "x₃"],
                axes = [1, 2, 3],
                args...)

SimplexPlots.simplex_scatter! — Function

simplex_scatter!(x::AbstractMatrix{Float64};
                 axes = [1, 2, 3],
                 args...)

Like simplex_scatter, but adds to an existing plot.

simplex_scatter!(x::Vector{Vector{Float64}};
                 axes = [1, 2, 3],
                 args...)

simplex_scatter!(x::Vector{Float64};
                 axes = [1, 2, 3],
                 args...)

SimplexPlots.simplex_histogram — Function

simplex_histogram(x::AbstractMatrix{Float64};
                  bins = 50,
                  vertices = ["x₁", "x₂", "x₃"],
                  axes = [1, 2, 3]
                  color = :jet,
                  args...)

Plots a simplex histogram of the vectors in x (row or column, depending on the orientation of the matrix x). The keyword arguments are:

bins: resolution of histogram (number of triangular bins along each edge of the simplex)

vertices: labels at the vertices of the simplex (set to nothing to suppress the labels)

axes: axis ordering

color: colorscheme for the color gradient

Further keyword arguments are passed to Plots.histogram.

simplex_histogram(x::Vector{Vector{Float64}};
                  bins = 50,
                  vertices = ["x₁", "x₂", "x₃"],
                  axes = [1, 2, 3],
                  color = :jet,
                  args...)

SimplexPlots.simplex_heatmap — Function

simplex_heatmap(x::AbstractMatrix{Float64};
                method = "square",
                vertices = ["x₁", "x₂", "x₃"],
                color = :jet,
                args...)

Plot a simplex heatmap, cell (i,j) in the array x supplying the value at the (i,j)th position.

Two methods for plotting a heatmap are available. The default (method = "square") fills the simplex with squares. This works well if the dimensionality of x is high, and is fast. The second method, (method = "triangle") fills the simplex with triangles. It is slow, particularly if the dimensionality of x is high.

Each method throws away values in x which correspond to combinations of coordinates which lie outside the simplex.

Vertex labels are controlled with the vertices keyword argument, color scheme with the color argument. Further keyword arguments are passed to Plots.heatmap.